sipo/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.1
Flask-Session==0.8.0
Flask-Migrate==4.1.0
Werkzeug==3.1.3
pandas==2.3.3
numpy==2.3.4
scipy==1.16.2
python-dotenv==1.1.1
bcrypt==5.0.0
openpyxl==3.1.5
PyJWT==2.10.1

# Testing dependencies
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Oracle Database

oracledb==2.5.1
ortools==9.11.4210
xlsxwriter==3.1.2
